feat: show prompt for restricted app when getting UDID
When trying to get the UDID of a restricted app, show a prompt to the user explaining why the UDID cannot be obtained.
