--- app/views/releases/body/_app_limited.html.slim
@@ -1,9 +1,0 @@
-.install-limited.d-none data-release-download-target="installLimited"
-  .cover
-    .app-limited-tips
-      span.text.text-right
-        = t('releases.show.app_touch_menu')
-      = image_tag('icons/short-arrow.png')
-
-  .alert.alert-danger.alert-dismissible
-    h5 = t('releases.show.app_limited')
