+++ app/views/shared/_install_limited.html.slim
@@ -1,0 +1,15 @@
+ruby:
+  attributes = {
+    "data-controller" => "install-limited",
+    "data-install-limited-keywords-value" => Setting.preset_install_limited.to_json,
+    "data-install-limited-target" => 'modal',
+    "data-install-limited-open-safari-value" => safari_text,
+    "data-install-limited-open-brower-value" => brower_text
+  }
+
+.install-limited.d-none *attributes
+  .cover
+    .app-limited-tips
+      span.text.text-right
+        = t('releases.show.app_touch_menu')
+      = image_tag('icons/short-arrow.png')
