+++ app/views/udid/index.html.slim
@@ -11,6 +11,8 @@
     data-udid-macos-tip-value="#{t('.macos_tip')}"
     data-udid-nonapple-tip-value="#{t('.nonapple_tip')}"
   ]
+    == render 'shared/install_limited', safari_text: t('.app_open_in_safari'), brower_text: t('.app_open_in_webbrower')
+
     .card.mb-3
       .card-body.text-center data-udid-target="qrcode"
         = install_qrcode_image_tag
