+++ config/locales/zealot/en.yml
@@ -438,7 +438,6 @@ en:
       api: API
     show:
       enter_password: Enter password
-      app_limited: Install limited in current app
       app_touch_menu: Click menu right-top corner
       app_open_in_safari: Choose "Open in Safari" and install
       app_open_in_webbrower: Open with any browser and install
@@ -785,7 +784,9 @@ en:
       apple_tip: Use iPhone or iPad to scan the QR Code to install
       macos_tip: System limited to detect macOS's chip. choose to use the iPhone and iPad scan code or press button to get the device UDID please.
       nonapple_tip: :udid.index.apple_tip
-      fetch_udid: Get your UDID now
+      fetch_udid: Fetch your UDID now
+      app_open_in_safari: Choose "Open in Safari" and fetch
+      app_open_in_webbrower: Open with any browser and fetch
       help:
         title: Why need to install profile to fetch your UDID?
         body_html: |
